+++ pysatNASA/instruments/icon_fuv.py
@@ -39,7 +39,7 @@ ICON_L27_Ion_Density becomes Ion_Density.  To retain the original names, use
 
 import datetime as dt
 import functools
-import warnings
+import numpy as np
 
 import pysat
 from pysat.instruments.methods import general as mm_gen
@@ -100,7 +100,7 @@ def clean(self):
 
     """
 
-    warnings.warn("Cleaning actions for ICON FUV are not yet defined.")
+    pysat.logger.warning("Cleaning actions for ICON FUV are not yet defined.")
     return
 
 
@@ -131,6 +131,42 @@ download = functools.partial(cdw.download, supported_tags=download_tags)
 # Set the list_remote_files routine
 list_remote_files = functools.partial(cdw.list_remote_files,
                                       supported_tags=download_tags)
+
+
+def filter_metadata(meta_dict):
+    """Filter FUV metadata to remove warnings during loading.
+
+    Parameters
+    ----------
+    meta_dict : dict
+        Dictionary of metadata from file
+
+    Returns
+    -------
+    dict
+        Filtered FUV metadata
+
+    """
+
+    vars = ['ICON_L24_UTC_Time', 'ICON_L25_Inversion_Method']
+    for var in vars:
+        if var in meta_dict:
+            meta_dict[var]['FillVal'] = np.nan
+
+    vars = ['ICON_L25_Start_Times', 'ICON_L25_Stop_Times', 'ICON_L25_UTC_Time']
+    for var in vars:
+        if var in meta_dict:
+            meta_dict[var]['FillVal'] = np.nan
+            meta_dict[var]['ValidMin'] = np.nan
+            meta_dict[var]['ValidMax'] = np.nan
+
+    # Deal with string arrays
+    for var in meta_dict.keys():
+        if 'Var_Notes' in meta_dict[var]:
+            meta_dict[var]['Var_Notes'] = ' '.join(pysat.utils.listify(
+                meta_dict[var]['Var_Notes']))
+
+    return meta_dict
 
 
 def load(fnames, tag=None, inst_id=None, keep_original_names=False):
@@ -179,7 +215,17 @@ def load(fnames, tag=None, inst_id=None, keep_original_names=False):
               'min_val': ('ValidMin', float),
               'max_val': ('ValidMax', float), 'fill_val': ('FillVal', float)}
 
-    data, meta = pysat.utils.load_netcdf4(fnames, epoch_name='Epoch',
-                                          pandas_format=pandas_format,
-                                          labels=labels)
+    meta_translation = {'FieldNam': 'plot', 'LablAxis': 'axis',
+                        'FIELDNAM': 'plot', 'LABLAXIS': 'axis',
+                        'Bin_Location': 'bin_loc'}
+
+    drop_labels = ['Valid_Max', 'Valid_Min', 'Valid_Range', 'SCALETYP',
+                   'TYPE', 'CONTENT']
+
+    data, meta = pysat.utils.io.load_netcdf(fnames, epoch_name='Epoch',
+                                            pandas_format=pandas_format,
+                                            labels=labels,
+                                            meta_processor=filter_metadata,
+                                            meta_translation=meta_translation,
+                                            drop_meta_labels=drop_labels)
     return data, meta